server/pyproject.toml
[project]
name = "gps-monitor"
version = "0.1.0"
description = "GPS 監視サーバ - ラズパイからGPSを取得しSlackへ通知する"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gps_monitor", "gps_web"]
